requirements.txt
asgiref==3.8.0
attrs==25.3.0
boto3==1.39.4
botocore==1.39.4
certifi==2025.7.14
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==5.0.3
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
s3transfer==0.13.0
six==1.17.0
sqlparse==0.4.4
stripe==12.3.0
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
